Started Client Rewrite
