+++ client/pom.xml
@@ -95,7 +95,11 @@
             <artifactId>Octree</artifactId>
             <version>master-104a60aa59-1</version>
         </dependency>
-
+        <dependency>
+            <groupId>org.jetbrains</groupId>
+            <artifactId>annotations</artifactId>
+            <version>19.0.0</version>
+        </dependency>
     </dependencies>
 
 </project>
