--- game/src/main/java/org/kakara/game/client/ClientWorldManager.java
@@ -1,13 +1,0 @@
-package org.kakara.game.client;
-
-import org.kakara.core.world.World;
-import org.kakara.core.world.WorldManager;
-
-import java.util.UUID;
-
-public class ClientWorldManager implements WorldManager {
-    @Override
-    public World getWorldByUUID(UUID uuid) {
-        return null;
-    }
-}
