--- game/src/main/java/org/kakara/game/client/SaveLoadException.java
@@ -1,5 +1,0 @@
-package org.kakara.game.client;
-
-//TODO move to core
-public class SaveLoadException extends Exception {
-}
